pm-daemon/requirements.txt
mcp
fastmcp
watchdog
requests
markdown
python-dotenv
